Fixed Time axis to always be in sync with incomoing data. Logging functionality almost implemented, timecode needs to be reformatted. Enforced thread safe plotting procedures. Added mouse coordinates (need to be thought through further to implement gridlines or a crosshair).  Refactored.

+++ PlotApp/Form1.cs
@@ -1,26 +1,29 @@
 using ScottPlot;
-using ScottPlot.WinForms;
+using System.Text;
 
 namespace PlotApp
 {
     public partial class PlotApp : Form
     {
-
-        double counter = DateTime.Now.ToNumber(); // program counter
-        double start = DateTime.Now.ToNumber(); // program starting time
-        readonly int seconds = 10; // number of previous seconds to be plotted
-        const int MaxPoints = 2000; // number of previous points to keep in memory
-        const int refresh_rate = 10; // refresh rate
+        static DateTime currentTime = DateTime.Now;
+        double counter = currentTime.ToNumber(); // program counter
+        double start = currentTime.ToNumber(); // program starting time
+        readonly int seconds = 10; // number of previous seconds to be plotted in n second view mode
+        const int MaxPoints = 600; // number of previous points to keep in memory
+        const int refresh_rate = 50; // refresh rate (1000/number)
         const int NumberOfSignals = 6; // signals chosen to analyze
-        List<double>[] data = new List<double>[NumberOfSignals]; // incoming data
+        List<double>[] data = Enumerable.Range(0, NumberOfSignals).Select(_ => new List<double>()).ToArray();
         const double conversion = 1000 / refresh_rate;
-        const double secondtoSample = (1D/86400)/conversion; // seconds to percent
+        const double secondtoSample = (1D / 86400) / conversion; // second to percent and convert based off of refresh rate
         int currentSampleCount = 0;
 
+
+        // make axis and signal arrays
         readonly ScottPlot.AxisPanels.LeftAxis[] axisL = new ScottPlot.AxisPanels.LeftAxis[NumberOfSignals];
         readonly ScottPlot.AxisPanels.RightAxis[] axisR = new ScottPlot.AxisPanels.RightAxis[NumberOfSignals];
         readonly ScottPlot.Plottables.Signal[] sig = new ScottPlot.Plottables.Signal[NumberOfSignals];
 
+        // define order of colors for traces and axes
         readonly ScottPlot.Color[] palette =
         [
             Colors.SeaGreen,
@@ -35,7 +38,7 @@ namespace PlotApp
             Colors.DarkKhaki,
         ];
 
-        // temporary button toggles
+        // button toggles
         bool view = false;
         bool pause = false;
         readonly bool[] trace = new bool[NumberOfSignals]; // active low
@@ -45,87 +48,67 @@ namespace PlotApp
         public PlotApp()
         {
             InitializeComponent();
-            InitializeArrays();
-            CreateChart();
-        }
-
-        private void InitializeArrays() // initialize data arrays to empty for appending
-        {
-            for (int i = 0; i < NumberOfSignals; i++)
-            {
-                data[i] = [];
-                trace[i] = false;
-            }
-        }
-
-        private void Form1_Load(object sender, EventArgs e) // enable timer
-        {
             // Create and enable timer in ms
-            System.Windows.Forms.Timer timer1 = new System.Windows.Forms.Timer
+            System.Timers.Timer timer1 = new System.Timers.Timer
             {
                 Enabled = true,
                 Interval = refresh_rate
             };
             timer1.Start();
-            timer1.Tick += timer1_Tick;
+            timer1.Elapsed += timer1_Tick;
+            CreateChart();
         }
 
+        // generate sinusoid value
         int c = 1;
         public double GetNextSineValue()
         {
             const int SamplesPerPeriod = 600;
-            // Increment the sample count
             currentSampleCount++;
-
-            // Reset the sample count if it exceeds the specified limit
             if (currentSampleCount > SamplesPerPeriod)
             {
                 currentSampleCount = 1;
                 c++;
             }
-
-            // Calculate the phase based on the current sample count
             double phase = 2 * Math.PI * currentSampleCount / SamplesPerPeriod;
-
-            // Calculate the sine value
             double sineValue = Math.Sin(phase);
-
-            // Return the same value 6 times in a row
-            return c+sineValue;
+            return c + sineValue;
         }
 
-        private void updateChart() // update array values and generate legend
+        private void updateData()
         {
-            // receive new data
+            // receive new data and remove old data if necessary
             double[] newValue = new double[NumberOfSignals];
 
             for (int i = 0; i < NumberOfSignals; i++)
             {
-                newValue[i] = GetNextSineValue();
-                if(i == 0)
+                newValue[i] = GetNextSineValue() + (rand.NextDouble()*0.2);
+                if (i == 0)
                 {
                     newValue[i] -= c;
                 }
-                newValue[i] *= (i+1);
+                newValue[i] *= (i + 1);
                 data[i] = [.. data[i], newValue[i]];
-            }
 
-            for (int i = 0; i<NumberOfSignals; i++)
-            {
-                if (data[i].Count > MaxPoints)
+                if (data[i].Count > MaxPoints && !pause)
                 {
                     data[i].RemoveRange(0, data[i].Count - MaxPoints);
                 }
             }
+            WriteDataToCsv();
+        }
 
-            // generate the signal
+        private void updateChart() // update array values and generate legend
+        {
+            // generate the signal assign axes and calculate time offset
             formsPlot1.Plot.Clear();
             for (int i = 0; i < NumberOfSignals; i++)
             {
                 if (!trace[i])
                 {
                     sig[i] = formsPlot1.Plot.Add.Signal(data[i], secondtoSample, color: palette[i]);
-                    if(i%2 == 0)
+                    
+                    if (i % 2 == 0)
                     {
                         sig[i].Axes.YAxis = axisL[i];
                     }
@@ -133,47 +116,72 @@ namespace PlotApp
                     {
                         sig[i].Axes.YAxis = axisR[i];
                     }
-                    sig[i].Data.XOffset = counter;
+
+                    if (data[i].Count >= MaxPoints)
+                    {
+                        sig[i].Data.XOffset = counter - (secondtoSample*MaxPoints);
+                    }
+                    else
+                    {
+                        sig[i].Data.XOffset = start;
+                    }
                 }
             }
 
             // increment x-axis
             dataX = [.. dataX, counter];
             counter += secondtoSample;
-
-            // Write data to CSV when the limit is reached
-            if (dataX.Count() == MaxPoints)
-            {
-                WriteDataToCsv();
-            }
-
         }
 
-        private void timer1_Tick(object sender, EventArgs e) // callback performed at tick rate
+    private void timer1_Tick(object sender, EventArgs e) // callback performed at tick rate
         {
+            updateData();
             // update chart and view mode
-            updateChart();
-
-            // stop refreshing if pause is toggled
-            if (!pause)
+            try
             {
-                formsPlot1.Refresh();
-                viewMode();
+                formsPlot1.Invoke((MethodInvoker)delegate {
+                    // force run on UI thread
+                    updateChart();
+                    if (!pause)
+                    {
+                        formsPlot1.Refresh();
+                        viewMode();
+
+                    }
+                    else
+                    {
+                        formsPlot1.MouseMove += (s, e) =>
+                        {
+                            Pixel mousePixel = new(e.X, e.Y);
+                            Coordinates mouseCoordinates = formsPlot1.Plot.GetCoordinates(mousePixel);
+                            Text = $"X={((mouseCoordinates.X) - start)*10000:N3}, Y={mouseCoordinates.Y:N3}";
+                        };
+
+                        formsPlot1.MouseDown += (s, e) =>
+                        {
+                            Pixel mousePixel = new(e.X, e.Y);
+                            Coordinates mouseCoordinates = formsPlot1.Plot.GetCoordinates(mousePixel);
+                            Text = $"X={((mouseCoordinates.X) - start)*10000:N3}, Y={mouseCoordinates.Y:N3} ";
+                        };
+                    }
+                });
             }
+            catch (Exception) { }
         }
 
         private void CreateChart() // set plot labels
         {
+            // remove excess borders
             formsPlot1.Plot.Axes.Remove(Edge.Left);
             formsPlot1.Plot.Axes.Remove(Edge.Right);
             formsPlot1.Plot.Axes.Remove(Edge.Top);
 
             // handle n amount of axes
-            for (int i = 0; i < NumberOfSignals;i++)
+            for (int i = 0; i < NumberOfSignals; i++)
             {
                 if (data[i] != null)
                 {
-                    if(i%2 == 0)
+                    if (i % 2 == 0)
                     {
                         axisL[i] = formsPlot1.Plot.Axes.AddLeftAxis();
                         axisL[i].Color(palette[i]);
@@ -185,7 +193,8 @@ namespace PlotApp
                     }
                 }
             }
-            formsPlot1.Plot.Axes.DateTimeTicks(Edge.Bottom);
+            // make bottom axes date and time and refresh chart
+            formsPlot1.Plot.Axes.DateTimeTicksBottom();
             formsPlot1.Refresh();
         }
 
@@ -194,21 +203,32 @@ namespace PlotApp
             if (view) // view last n
             {
                 formsPlot1.Plot.Axes.AutoScale();
-                formsPlot1.Plot.Axes.SetLimitsX(counter-(secondtoSample*seconds*conversion)+0.000235, counter+0.000235);
-                //double max = 0;
-                //for(int i = 0;i < NumberOfSignals; i++)
-                //{
-                //    if (data[i].Max() > max)
-                //    {
-                //        max = data[i].Max();
-                //    }
-                //}
-                //formsPlot1.Plot.Axes.SetLimitsY(0, max);
+                formsPlot1.Plot.Axes.SetLimitsX(counter - (secondtoSample * seconds * conversion), counter);
             }
             else // auto scale
             {
                 formsPlot1.Plot.Axes.AutoScale();
             }
+        }
+
+        private void WriteDataToCsv()
+        {
+            StringBuilder csvContent = new StringBuilder();
+            csvContent.AppendLine("Time," + string.Join(",", Enumerable.Range(1, NumberOfSignals).Select(i => $"Signal{i}")));
+
+            for (int j = 0; j < dataX.Length; j++)
+            {
+                DateTime timestamp = currentTime.AddSeconds(dataX[j]);
+                string formattedTimestamp = timestamp.ToString("yyyy-MM-dd HH:mm:ss.fff");
+                csvContent.AppendLine($"{formattedTimestamp},{string.Join(",", Enumerable.Range(0, NumberOfSignals).Select(i => data[i].Count > j ? data[i][j].ToString() : ""))}");
+            }
+
+            try
+            {
+                using StreamWriter writer = new StreamWriter("data.csv");
+                writer.Write(csvContent.ToString());
+            }
+            catch (Exception){}
         }
 
         Random rand = new();
@@ -261,7 +281,7 @@ namespace PlotApp
         {
             int num = 5;
             traceException(num);
-        } 
+        }
         private void traceException(int num)
         {
             try
@@ -270,26 +290,5 @@ namespace PlotApp
             }
             catch (Exception) { }
         }
-
-        private void WriteDataToCsv()
-        {
-            try
-            {
-                using StreamWriter writer = new("data.csv");
-                // Write header
-                writer.WriteLine("Time," + string.Join(",", Enumerable.Range(1, NumberOfSignals).Select(i => $"Signal{i}")));
-
-                // Write data
-                for (int j = 0; j < dataX.Count(); j++)
-                {
-                    writer.WriteLine($"{dataX[j]},{string.Join(",", Enumerable.Range(0, NumberOfSignals).Select(i => data[i].Count > j ? data[i][j].ToString() : ""))}");
-                }
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show($"Error writing to CSV: {ex.Message}");
-            }
-        }
-
     }
 }
